debug refactor all is ok, skip autocommit case at present
