+++ behave_dble/features/cluster/zk_sequence.feature
@@ -104,10 +104,10 @@ Feature: Functional testing of global sequence generated by distributed timestam
     Given sleep "60" seconds
     Then Start dble in "dble-1"
     Then execute sql in "dble-1" in "user" mode
-      | user | passwd | conn   | toClose  | sql                                                      | expect         | db     |
-      | test | 111111 | conn_0 | True     |drop table if exists test_auto                        | success        | schema1 |
-      | test | 111111 | conn_0 | True     |create table test_auto(id bigint,time char(120))    | success        | schema1 |
-      | test | 111111 | conn_0 | True     |insert into test_auto values(1)                       | Global sequence has reach to max limit and can generate duplicate sequences        | schema1 |
+      | conn   | toClose  | sql                                                | expect       | db      |
+      | conn_0 | False    |drop table if exists test_auto                      | success      | schema1 |
+      | conn_0 | False    |create table test_auto(id bigint,time char(120))    | success      | schema1 |
+      | conn_0 | True     |insert into test_auto values(1)                     | Global sequence has reach to max limit and can generate duplicate sequences  | schema1 |
     Given update file content "{install_dir}/dble/conf/sequence_distributed_conf.properties" in "dble-1" with sed cmds
     """
     /START_TIME/c START_TIME=2010-11-04 09:42:54
@@ -116,5 +116,5 @@ Feature: Functional testing of global sequence generated by distributed timestam
     Given sleep "60" seconds
     Then Start dble in "dble-1"
     Then execute sql in "dble-1" in "user" mode
-      | user | passwd | conn   | toClose  | sql                                                      | expect         | db     |
-      | test | 111111 | conn_0 | True     |insert into test_auto values(1)                       | success        | schema1 |
+      | sql                              | expect         | db     |
+      |insert into test_auto values(1)   | success        | schema1 |
