+++ behave_dble/features/install_uninstall/single_dble_and_zk_cluster.feature
@@ -40,17 +40,3 @@ Feature: install dble cluster, degrade to single dble, recover to cluster
     dble-3
     """
     Given stop dble cluster and zk service
-
-  Scenario: verify dnindex.properties should not be empty after restart dble #3
-    Given stop dble cluster and zk service
-    Given a clean environment in all dble nodes
-    Given install dble in all dble nodes
-    Given replace config files in all dbles with command line config
-    Given config zookeeper cluster in all dble nodes with "all zookeeper hosts"
-    Given reset dble registered nodes in zk
-    Given Restart dble in "dble-1" success
-    Then check following text exist "Y" in file "/opt/dble/conf/dnindex.properties" in host "dble-1"
-    """
-    localhost1=0
-    """
-    Given stop dble cluster and zk service
