+++ behave_dble/features/steps/dble_steps.py
@@ -10,7 +10,7 @@ import random
 import time
 import MySQLdb
 
-from MySQLSteps import *
+from mysql_steps import *
 from lib.DBUtil import DBUtil
 from behave import *
 from hamcrest import *
@@ -188,3 +188,16 @@ def step_impl(context, host_name, result_key, mode_name=None):
 
     setattr(context, result_key, res)
 
+@Then('execute admin cmd  in "{host}" at background')
+def step_impl(context, host):
+    node = get_node(host)
+    sshClient = node.ssh_conn
+
+    context.logger.debug("btrace is running, start query!!!")
+    time.sleep(5)
+    for row in context.table:
+        query_meta = QueryMeta(row.as_dict(), "admin", node)
+
+        cmd = u"nohup mysql -u{} -p{} -P{} -c -e\"{}\" >/tmp/dble_query.log 2>&1 &".format(query_meta.user, query_meta.passwd, query_meta.port, query_meta.sql)
+        rc, sto, ste = sshClient.exec_command(cmd)
+        assert len(ste) == 0, "impossible err occur"
