--- behave_dble/features/steps/restart.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright (C) 2016-2020 ActionTech.
-# License: https://www.mozilla.org/en-US/MPL/2.0 MPL version 2 or higher.
-# @Time    : 2018/5/18 PM4:58
-# @Author  : zhaohongjie@actionsky.com
-
-
-
